requirements.txt
Pillow
mss
pywin32
